requirements.txt
PyQt6==6.7.1
Pillow==10.4.0
torch==2.3.1
transformers==4.43.3
onnxruntime==1.18.1
huggingface_hub==0.24.5
einops==0.8.0
timm==1.0.8
numpy==1.26.4
